+++ src/MathOperations.java
@@ -1,14 +1,12 @@
 public class MathOperations {
-    public static int add(int x, int y) {
-        return x + y;
-    }
-    public static int subtract(int x, int y) {
+    public static double add(double x, double y) {return x+y;}
+    public static double subtract(double x, double y) {
         return x - y;
     }
-    public static int multiply(int x, int y) {
+    public static double multiply(double x, double y) {
         return x*y;
     }
-    public static float divide(int x, int y) {
+    public static double divide(double x, double y) {
         if (y != 0) {
             return (float) x / y;
         } else {
